Refactor into smaller presets
